pyproject.toml
[project]
name = "Aegis"
version = "1.0.0"
description = "Aegis a LLM Red Team Platform"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.3.0",
    "textual>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "litellm>=1.0.0",
]

[project.scripts]
redteam = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
